# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/interviews": {
      "post": {
        "operationId": "InterviewController_create",
        "summary": "Create interview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewWithCandidateLinkResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      },
      "get": {
        "operationId": "InterviewController_findAll",
        "summary": "List interviews",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      }
    },
    "/interviews/{id}": {
      "get": {
        "operationId": "InterviewController_findOne",
        "summary": "Get interview by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      }
    },
    "/interviews/{id}/candidate-link": {
      "post": {
        "operationId": "InterviewController_generateCandidateLink",
        "summary": "Generate candidate interview link",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateLinkResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      }
    },
    "/interviews/{id}/complete": {
      "patch": {
        "operationId": "InterviewController_complete",
        "summary": "Complete interview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      }
    },
    "/interviews/{id}/validate": {
      "post": {
        "operationId": "InterviewController_validateAllAnswers",
        "summary": "Start validation for all submitted answers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartAllAnswerValidationsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      }
    },
    "/interviews/{id}/questions/{questionIndex}/validate": {
      "post": {
        "operationId": "InterviewController_validateAnswer",
        "summary": "Start validation for single answer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionIndex",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartAnswerValidationResultDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      }
    },
    "/interviews/{id}/results": {
      "get": {
        "operationId": "InterviewController_getResults",
        "summary": "Get interview results",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResultResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "Interview App API",
    "description": "HTTP API contract for Interview App backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "QuestionExpectedConceptDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "weight": {
            "type": "number"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "weight",
          "description"
        ]
      },
      "QuestionRedFlagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ]
          }
        },
        "required": [
          "id",
          "label",
          "severity"
        ]
      },
      "QuestionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "focus": {
            "type": "string"
          },
          "outputLanguage": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "subcategory": {
            "type": "string"
          },
          "questionText": {
            "type": "string"
          },
          "followUpQuestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expectedConcepts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionExpectedConceptDto"
            }
          },
          "redFlags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionRedFlagDto"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "weight": {
            "type": "number"
          },
          "sampleGoodAnswer": {
            "type": "string"
          },
          "minimumPassScore": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "outputLanguage",
          "questionText",
          "followUpQuestions",
          "expectedConcepts",
          "redFlags",
          "difficulty",
          "weight",
          "minimumPassScore",
          "tags",
          "metadata",
          "createdAt",
          "updatedAt",
          "deleted"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Validation failed"
          },
          "path": {
            "type": "string",
            "example": "/questions/invalid-id"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "CreateInterviewDto": {
        "type": "object",
        "properties": {
          "candidateName": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "questionIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "candidateName",
          "position",
          "questionIds"
        ]
      },
      "MediaArtifactDto": {
        "type": "object",
        "properties": {
          "mediaKey": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "fileSizeBytes": {
            "type": "number"
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "mediaKey",
          "contentType",
          "uploadedAt"
        ]
      },
      "AnswerBehaviorSignalsDto": {
        "type": "object",
        "properties": {
          "tabHiddenCount": {
            "type": "number"
          },
          "windowBlurCount": {
            "type": "number"
          },
          "pasteCount": {
            "type": "number"
          },
          "keydownCount": {
            "type": "number"
          },
          "resizeCount": {
            "type": "number"
          }
        },
        "required": [
          "tabHiddenCount",
          "windowBlurCount",
          "pasteCount",
          "keydownCount",
          "resizeCount"
        ]
      },
      "AnswerTranscriptDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "generatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isFinal": {
            "type": "boolean"
          }
        }
      },
      "AnswerEvaluationDto": {
        "type": "object",
        "properties": {
          "overallScore": {
            "type": "number"
          },
          "categoryScores": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          },
          "coveredConceptIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "missedConceptIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "redFlagIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "behaviorRisk": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ]
          },
          "summary": {
            "type": "string"
          },
          "decisionHint": {
            "type": "string",
            "enum": [
              "pass",
              "review",
              "fail"
            ]
          },
          "evaluatedAt": {
            "format": "date-time",
            "type": "string"
          }
        }
      },
      "AnswerValidationDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "idle",
              "queued",
              "processing",
              "completed",
              "failed"
            ]
          },
          "executionArn": {
            "type": "string"
          },
          "sourceVersionNumber": {
            "type": "number"
          },
          "requestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "AnswerDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "questionId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "recording",
              "submitted"
            ]
          },
          "mediaKey": {
            "type": "string"
          },
          "screenMediaKey": {
            "type": "string"
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string"
          },
          "durationSeconds": {
            "type": "number"
          },
          "retakeCount": {
            "type": "number"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string"
          },
          "camera": {
            "$ref": "#/components/schemas/MediaArtifactDto"
          },
          "screen": {
            "$ref": "#/components/schemas/MediaArtifactDto"
          },
          "behaviorSignals": {
            "$ref": "#/components/schemas/AnswerBehaviorSignalsDto"
          },
          "selectedVersionNumber": {
            "type": "number"
          },
          "transcript": {
            "$ref": "#/components/schemas/AnswerTranscriptDto"
          },
          "evaluation": {
            "$ref": "#/components/schemas/AnswerEvaluationDto"
          },
          "validation": {
            "$ref": "#/components/schemas/AnswerValidationDto"
          }
        },
        "required": [
          "questionIndex",
          "questionId",
          "status",
          "mediaKey",
          "uploadedAt"
        ]
      },
      "InterviewBehaviorSummaryDto": {
        "type": "object",
        "properties": {
          "riskLevel": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ]
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "notes"
        ]
      },
      "InterviewQuestionResultDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "questionId": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "categoryScores": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          },
          "summary": {
            "type": "string"
          },
          "decisionHint": {
            "type": "string",
            "enum": [
              "pass",
              "review",
              "fail"
            ]
          }
        },
        "required": [
          "questionIndex",
          "questionId"
        ]
      },
      "InterviewResultResponseDto": {
        "type": "object",
        "properties": {
          "overallScore": {
            "type": "number"
          },
          "summary": {
            "type": "string"
          },
          "categoryScores": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          },
          "rubricVersion": {
            "type": "string"
          },
          "decision": {
            "type": "string",
            "enum": [
              "proceed",
              "review",
              "reject"
            ]
          },
          "trustScore": {
            "type": "number"
          },
          "trustFlags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "behaviorSummary": {
            "$ref": "#/components/schemas/InterviewBehaviorSummaryDto"
          },
          "questionResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewQuestionResultDto"
            }
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "overallScore",
          "summary",
          "categoryScores",
          "completedAt"
        ]
      },
      "InterviewWithCandidateLinkResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "candidateName": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResponseDto"
            }
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerDto"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "processing",
              "completed",
              "failed"
            ]
          },
          "result": {
            "$ref": "#/components/schemas/InterviewResultResponseDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "candidateLink": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "candidateName",
          "position",
          "questions",
          "answers",
          "status",
          "createdAt",
          "updatedAt",
          "candidateLink"
        ]
      },
      "InterviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "candidateName": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResponseDto"
            }
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerDto"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "processing",
              "completed",
              "failed"
            ]
          },
          "result": {
            "$ref": "#/components/schemas/InterviewResultResponseDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "candidateName",
          "position",
          "questions",
          "answers",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CandidateLinkResponseDto": {
        "type": "object",
        "properties": {
          "candidateLink": {
            "type": "string"
          }
        },
        "required": [
          "candidateLink"
        ]
      },
      "StartAnswerValidationResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "idle",
              "queued",
              "processing",
              "completed",
              "failed"
            ]
          },
          "questionIndex": {
            "type": "number"
          },
          "sourceVersionNumber": {
            "type": "number"
          },
          "reused": {
            "type": "boolean"
          }
        },
        "required": [
          "status",
          "questionIndex",
          "sourceVersionNumber",
          "reused"
        ]
      },
      "StartAllAnswerValidationsResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "interviewId": {
            "type": "string"
          },
          "requestedCount": {
            "type": "number"
          },
          "queuedCount": {
            "type": "number"
          },
          "reusedCount": {
            "type": "number"
          },
          "skippedCount": {
            "type": "number"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StartAnswerValidationResultDto"
            }
          }
        },
        "required": [
          "ok",
          "interviewId",
          "requestedCount",
          "queuedCount",
          "reusedCount",
          "skippedCount",
          "answers"
        ]
      },
...
    }
  }
}